feat(rendercontenttype): include "sys" with content type ID in the generated interface (#4)
It's helpful to be able to inspect an entry's `sys.contentType.sys.id` field to see what type it is,
especially when a link can be one of many types. This makes the types support such a check.

BREAKING CHANGE: Users who use one content type's interface to accept another content type in a
function (for example, if one content type is a superset of another one) will need to use a union
type and allow all possible content types now.