feat(flag): add a "fields-only" mode for validating responses
At Intercom, we have a need to validate Contentful responses to ensure they match the schema
constructed by the codegen. This will prevent our marketing site from building if Contentful content
is unexpectedly unpublished, the validations are messed up, or (hopefully) any other discrepancy.
This flag is intended to support the use case where you want to ensure that all of the fields in the
Contentful response are well-formed, but don't want to validate auxilliary things like Sys, Assets,
etc.