fix(cli): return a non-zero exit code on failure
Previously, this script would silently fail when an error was encountered. This breaks builds. Now,
it will return an exit code of 1 when an error is encountered.

BREAKING CHANGE: This may lead to builds that previously succeeded breaking. This would mean that
there was previously an uncaught error in your setup. Well-configured, healthy builds should remain
unaffected.